plugins/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plugins"
version = "0.0.1"

dependencies = ["apache-airflow>=2.3"]
[project.entry-points.'airflow.policy']
_ = 'policies'